requirements.txt
astroid==1.5.3
colorama==0.3.9
Django==1.11.7
django-webpack-loader==0.5.0
djangorestframework==3.7.3
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
pylint==1.7.4
pytz==2017.3
six==1.11.0
wrapt==1.10.11
